mcp-server/pyproject.toml
[project]
name = "gitbeat-mcp-server"
version = "0.1.0"
description = "MCP Server for GitBeat music generation with ElevenLabs"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.uv]
dev-dependencies = []

[project.scripts]
gitbeat-mcp = "gitbeat_mcp.server:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
